\name{addEnvironment}
\alias{addEnvironment}
\title{Add a new environment}
\usage{
addEnvironment(api, data)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{data}{the environment in list format}
}
\description{
Post a new environment to the database
}
\details{
Requires authentication
}

