\name{addInteraction}
\alias{addInteraction}
\title{Add a new interaction}
\usage{
addInteraction(api, data)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{data}{the interaction in list format}
}
\description{
Post a new interaction to the database
}
\details{
Requires authentication
}

