% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addPopulation}
\alias{addPopulation}
\title{Add a new population}
\usage{
addPopulation(api, data)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{data}{the population in list format}
}
\description{
Post a new population to the database
}
\details{
Requires authentication
}

