% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addReference}
\alias{addReference}
\title{Add a new reference}
\usage{
addReference(api, data)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{data}{the reference in list format}
}
\description{
Post a new reference to the database
}
\details{
Requires authentication
}

