% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addTrait}
\alias{addTrait}
\title{Add a new trait}
\usage{
addTrait(api, data)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{data}{the trait in list format}
}
\description{
Post a new trait to the database
}
\details{
Requires authentication
}

