% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{as.data.frame.mangalGetResp}
\alias{as.data.frame.mangalGetResp}
\title{Coerce mangalGetResp to data.frame}
\usage{
\method{as.data.frame}{mangalGetResp}(responses = responses, ...)
}
\arguments{
\item{responses}{mangalGetResp class}
}
\description{
Coerce mangalGetResp to data.frame
}
