% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{as.data.frame.mgGetResponses}
\alias{as.data.frame.mgGetResponses}
\title{Coerce mgGetResponses to data.frame}
\usage{
\method{as.data.frame}{mgGetResponses}(responses)
}
\arguments{
\item{responses}{mgGetResponses class}
}
\description{
Coerce mgGetResponses to data.frame
}
