% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{cl_to_sf}
\alias{cl_to_sf}
\alias{cl_to_sf.coleoGetResp}
\alias{cl_to_sf.data.frame}
\title{Fonction générique pour convertir les sorties de l'API en objet de classe spatiale (sf)}
\usage{
cl_to_sf(data)

\method{cl_to_sf}{coleoGetResp}(data)

\method{cl_to_sf}{data.frame}(data)
}
\arguments{
\item{data}{`list` de la classe `coleoGetResp` retourné par les fonctions GET du package coléo (ex. [get_sites()]) ou `data.frame` ([as.data.frame]) , voir @details}
}
\value{
Retourne un objet de la classe `sf::st_sf`
}
\description{
Fonction générique pour convertir les sorties de l'API en objet de classe spatiale (sf)
}
\details{
L'objet d'entrée doit contenir les colonnes `geom.type` & `geom.coordinates`
}
\examples{
\dontrun{
plot(cl_to_sf(get_sites()))
} 
}
\seealso{
[sf_cells()], [sf_sites()]
}
