% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear memoise cache}
\usage{
clear_cache()
}
\value{
logical
}
\description{
Clear memoise cache
}
\details{
Clear memoise cache generated by \code{mem_get()}, a memoise function around \code{httr::GET} used in \code{get_gen} in \code{get_singleton}
}
