\name{collateGraph}
\alias{collateGraph}
\title{Collate graph}
\usage{
  collateGraph(API, dataset, network, ...)
}
\arguments{
  \item{API}{An \code{\link{mangalAPI}} object}

  \item{dataset}{A \code{list} object with the dataset
  information}

  \item{network}{A \code{list} object with the network
  information}
}
\description{
  Returns a network as an \code{\link{igraph}} object
}

