% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mgNetworks.R
\name{combine_mgNetworks}
\alias{combine_mgNetworks}
\title{Combine Mangal networks.}
\usage{
combine_mgNetworks(...)
}
\arguments{
\item{...}{objects of class \code{mgNetworksCollection} or \code{mgNetwork} or a list #' of objects of these classes.}
}
\value{
An object of class \code{mgNetworksCollection}
}
\description{
Combine \code{mgNetworksCollection} and \code{mgNetwork} objects into a
\code{mgNetworksCollection} object.
}
\examples{
mg_19 <- get_collection(search_networks(list(dataset_id = 19)))
mg_lagoon <- get_collection(search_datasets(query='lagoon\%'))
combine_mgNetworks(mg_19, mg_lagoon)

}
