% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELETE_network.R
\name{DELETE_network}
\alias{DELETE_network}
\title{Remove a network}
\usage{
DELETE_network(id, name)
}
\arguments{
\item{id}{A numeric, is the id of the network to remove}

\item{name}{A character, is the name of the network to remove
Users can only remove their own datasets}
}
\value{
The status and information of the resquest
}
\description{
Remove a network and associated information from Mangal. Delete on cascade
}
\author{
Gabriel Bergeron
}
\keyword{database}
