% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collection.R
\name{get_collection}
\alias{get_collection}
\title{Retrieve mangal networks collection}
\usage{
get_collection(x, ...)
}
\arguments{
\item{x}{\code{numeric} vector of mangal network IDs or object return by functions \code{search_*}}

\item{...}{arguments from \code{\link[rmangal:get_network_by_id]{rmangal::get_network_by_id()}}}
}
\value{
\itemize{
\item networks: \code{list} of object \code{mgNetwork} (see \code{\link[rmangal:get_network_by_id]{rmangal::get_network_by_id()}})
}
}
\description{
Retrieve mangal networks collection
}
\examples{
get_collection(c(1035:1037))
get_collection(search_networks(query="insect\%"))
get_collection(search_datasets(query="lagoon\%"))
}
