% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_fkey.R
\name{GET_fkey}
\alias{GET_fkey}
\title{GET id key from a Mangal entry}
\usage{
GET_fkey(table, attribute, value)
}
\arguments{
\item{table}{A element, must be the name of the targeted table with ""}

\item{attribute}{A element, must be the name of the targeted attribute with ""}

\item{value}{A element, must be the value of the targeted attribute with ""}
}
\value{
The value of the primary key of a specified entry
}
\description{
GET primary key of a table entry specified by the user
}
\author{
Gabriel Bergeron
}
\keyword{database}
