% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_from_fkey}
\alias{get_from_fkey}
\title{Get entries based on foreign key}
\usage{
get_from_fkey(endpoint, verbose = TRUE, ...)
}
\arguments{
\item{endpoint}{\code{character} API entry point}

\item{...}{foreign key column name with the id}
}
\value{
Object returned by \code{\link[rmangal:get_gen]{rmangal::get_gen()}}
}
\description{
Get entries based on foreign key
}
\details{
See endpoints available with \code{endpoints()}
}
\keyword{internal}
