% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_gen}
\alias{get_gen}
\title{GET generic function}
\usage{
get_gen(endpoint = NULL, query = NULL, limit = 100, flatten = TRUE,
  output = "data.frame", ...)
}
\arguments{
\item{endpoint}{`character` API entry point}

\item{query}{`list` list of params passed to the API}

\item{limit}{`integer` number of entries return by the API (max: 1000)}

\item{flatten}{`logical` flatten nested data.frame, see [jsonlite::flatten()]}

\item{output}{`character` output type: `data.frame`, `list`, `json`}

\item{...}{httr options, see [httr::GET()]}
}
\value{
Object of class `getSuccess` or `getError`.
`getSuccess` is a list with the body [httr::content()] and the server response [httr::response()]. 
`getError` has the exact same structure with the body empty.
}
\description{
GET generic function
}
\details{
See endpoints available with `print(endpoints)`
}
