\name{get_network}
\alias{get_network}
\title{Get a network}
\usage{
  get_network(api, id, level = "taxa", ...)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{id}{the id of the network}

  \item{level}{the level of aggregation}

  \item{...}{reserved for future options}
}
\description{
  ...
}

