\name{getElements}
\alias{getElements}
\title{Get aggregated informations on a network}
\usage{
getElements(api, id, level = "taxa", ...)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{id}{the id of the network}

  \item{level}{the level of aggregation}

  \item{...}{reserved for future options}
}
\description{
Internal use
}

