% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEnvironment}
\alias{getEnvironment}
\title{Get an environmental measure}
\usage{
getEnvironment(api, id)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{id}{the identifier of the environment}
}
\description{
Returns a list of environmental measures
}

