\name{getInteraction}
\alias{getInteraction}
\title{Get a particular interaction}
\usage{
  getInteraction(API, dataset, network, id)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{dataset}{A \code{list} object returned by
  \code{\link{getDataset}}}

  \item{network}{A \code{list} object returned by
  \code{\link{getNetwork}}}

  \item{id}{The \code{id} of the interaction to return}
}
\description{
  Returns an interaction from a network object
}

