\name{getNetwork}
\alias{getNetwork}
\title{Get a particular network}
\usage{
  getNetwork(API, dataset, id)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{dataset}{A \code{list} object returned by
  \code{\link{getDataset}}}

  \item{id}{The \code{id} of the network to return}
}
\description{
  Returns a network object from a dataset
}

