\name{getPop}
\alias{getPop}
\title{Get a specific population}
\usage{
  getPop(API, taxa, id)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{id}{A string giving the id of the population}

  \item{taxa}{The taxa object to which the population
  belongs}
}
\description{
  Return a list with the informations on a specific
  population
}

