\name{getPopstate}
\alias{getPopstate}
\title{Get a particular population state}
\usage{
  getPopstate(API, taxa, pop, id)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{taxa}{A \code{taxa} object in \code{list} format}

  \item{pop}{A \code{population} object in \code{list}
  format}

  \item{id}{The ID of the population state to retrieve}
}
\description{
  Returns the state of a given population
}

