% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{is.mangalGetResp}
\alias{is.mangalGetResp}
\title{Test on class `coleoGetResp` (Objet S3)}
\usage{
is.mangalGetResp(x)
}
\arguments{
\item{x}{Objet à tester}
}
\description{
Test on class `coleoGetResp` (Objet S3)
}
