% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{is.mangalPostResp}
\alias{is.mangalPostResp}
\title{Test on class `coleoPostResp` (Objet S3)}
\usage{
is.mangalPostResp(x)
}
\arguments{
\item{x}{Objet à tester}
}
\description{
Test on class `coleoPostResp` (Objet S3)
}
