% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{is.mgGetResponses}
\alias{is.mgGetResponses}
\title{Test on class \code{coleoGetResponses} (Objet S3)}
\usage{
is.mgGetResponses(x)
}
\arguments{
\item{x}{R object to test}
}
\description{
Test on class \code{coleoGetResponses} (Objet S3)
}
