% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_list.R
\name{json_list}
\alias{json_list}
\title{Dataframe into json list}
\usage{
json_list(df)
}
\arguments{
\item{df}{A dataframe, header must be the names of attributes}
}
\value{
A list of json data, each level of the list is ready for injection in Mangal
}
\description{
Convert a dataframe into a list of json data
}
\author{
Gabriel Bergeron
}
\keyword{manip}
