% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List of datasets with networks and reference attached}
\usage{
list_datasets(search, ...)
}
\arguments{
\item{search}{\code{character} keyword to search (case sensitive)}

\item{...}{arguments from \code{\link[rmangal:get_gen]{rmangal::get_gen()}}}
}
\value{
\code{data.frame} of datasets with all networks and the original reference.
}
\description{
List of datasets with networks and reference attached
}
\details{
See endpoints available with \code{print(endpoints)}
}
\examples{
listDatasets()
}
