% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listDataset}
\alias{listDataset}
\title{List the datasets currently in the DB}
\usage{
listDataset(api, ...)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{...}{additional parameters (filters) to be passed to \code{\link{mangalList}}}
}
\description{
Returns a list of datasets
}

