% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDatasets.R
\name{listDatasets}
\alias{listDatasets}
\title{List of datasets with networks and reference attached}
\usage{
listDatasets(...)
}
\arguments{
\item{...}{arguments from \code{\link[rmangal:get_gen]{rmangal::get_gen()}}}
}
\value{
\code{data.frame} of datasets with all networks and the original reference.
}
\description{
List of datasets with networks and reference attached
}
\details{
See endpoints available with \code{print(endpoints)}
}
\examples{
listDatasets()
}
