\name{listEnvironment}
\alias{listEnvironment}
\title{List the environments}
\usage{
listEnvironment(api, ...)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{...}{additional parameters (filters) to be passed
  to \code{\link{mangalList}}}
}
\description{
Returns a list of environments
}

