% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listInteraction}
\alias{listInteraction}
\title{List the interactions currently in the DB}
\usage{
listInteraction(api, ...)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{...}{additional parameters (filters) to be passed to \code{\link{mangalList}}}
}
\description{
Returns a list of interactions
}

