\name{listInteractions}
\alias{listInteractions}
\title{Lists the interactions within a network}
\usage{
  listInteractions(API, dataset, network)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{dataset}{A \code{list} object representing a
  dataset}

  \item{network}{A \code{list} object representing a
  network}
}
\description{
  Makes a list of all interactions part of a network
}

