\name{listNetworks}
\alias{listNetworks}
\title{Lists the networks that are part of a dataset}
\usage{
  listNetworks(API, dataset)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{dataset}{A \code{list} object representing a
  dataset}
}
\description{
  Makes a list of all networks part of a dataset
}

