\name{listPop}
\alias{listPop}
\title{Lists the populations of a particular taxa}
\usage{
  listPop(API, taxa)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{taxa}{A \code{taxa} object in \code{list} format}
}
\description{
  Makes a list of all populations associated to a
  particular taxa.
}

