\name{listPopstate}
\alias{listPopstate}
\title{Lists the states of a particular population}
\usage{
  listPopstate(API, taxa, pop)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{taxa}{A \code{taxa} object in \code{list} format}

  \item{pop}{A \code{population} object in \code{list}
  format}
}
\description{
  Makes a list of all states associated to a particular
  population.
}

