\name{makefilter}
\alias{makefilter}
\title{Construct a filtering query}
\usage{
makefilter(api, type, filter)
}
\arguments{
  \item{api}{the api to connect to}

  \item{type}{the type of resource to filter}

  \item{filter}{a list with objects field, relation,
  target}
}
\description{
Returns a string formatted as a filtering query
}

