% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeTaxaTable}
\alias{makeTaxaTable}
\title{Make a table with taxa level infos}
\usage{
makeTaxaTable(api, id)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{id}{a vector of taxa id}
}
\description{
Make a table with taxa level infos
}

