% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangal.R
\docType{package}
\name{mangal}
\alias{mangal}
\alias{mangal-package}
\title{mangal: A package to ....}
\description{
The description of the package
}
\section{mangal functions}{

  Describe mangal function here
}

