\name{mangalGet}
\alias{mangalGet}
\title{Get an object of a given type

Returns an object}
\usage{
mangalGet(api, type, id = NULL)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{type}{a type of object}

  \item{id}{the ID of the object}
}
\description{
Get an object of a given type

Returns an object
}

