% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mangalList}
\alias{mangalList}
\title{List all objects of a given type}
\usage{
mangalList(api, type, filtering = NULL)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{type}{a type of object}

\item{filtering}{a vector of filters to restrict the returned objects}
}
\description{
Returns a list of objects of a given type
}

