% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mangalPatch}
\alias{mangalPatch}
\title{Patch an object

Patch an existing object}
\usage{
mangalPatch(api, type, data)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{type}{the type of object to patch}

\item{data}{the object in list form}
}
\description{
Patch an object

Patch an existing object
}

