% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{mg_to_sf}
\alias{mg_to_sf}
\title{Coerce body return by the API to an sf object}
\usage{
mg_to_sf(body)
}
\arguments{
\item{body}{\code{data.frame} return by the API call}
}
\value{
sf object
}
\description{
Coerce body return by the API to an sf object
}
