% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.igraph.R
\name{mgNetworksCollection.as.igraph}
\alias{mgNetworksCollection.as.igraph}
\title{Coerce mgNetworksCollection to igraph object}
\usage{
mgNetworksCollection.as.igraph()
}
\description{
Coerce mgNetworksCollection to igraph object
}
