% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pagerResources}
\alias{pagerResources}
\title{Read resources}
\usage{
pagerResources(api, queryset)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{queryset}{the entry point in the API}
}
\description{
Read a page list of resources
}

