% Generated by roxygen2 (4.0.1): do not edit by hand
\name{patchDataset}
\alias{patchDataset}
\title{Patch a dataset}
\usage{
patchDataset(api, data)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{data}{the dataset in list format}
}
\description{
Patch a dataset from the database
}
\details{
Requires authentication
}

