\name{patchInteraction}
\alias{patchInteraction}
\title{Patch an interaction}
\usage{
patchInteraction(api, data)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{data}{the interaction in list format}
}
\description{
Patch an interaction from the database
}
\details{
Requires authentication
}

