% Generated by roxygen2 (4.0.1): do not edit by hand
\name{patchNetwork}
\alias{patchNetwork}
\title{Patch a network}
\usage{
patchNetwork(api, data)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{data}{the network in list format}
}
\description{
Modify the informations for a network
}
\details{
Requires authentication
}

