% Generated by roxygen2 (4.0.1): do not edit by hand
\name{patchPopulation}
\alias{patchPopulation}
\title{Patch a population}
\usage{
patchPopulation(api, data)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{data}{the population in list format}
}
\description{
Modify the informations for a population
}
\details{
Requires authentication
}

