% Generated by roxygen2 (4.0.1): do not edit by hand
\name{patchTrait}
\alias{patchTrait}
\title{Patch a trait}
\usage{
patchTrait(api, data)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{data}{the trait in list format}
}
\description{
Modify the informations for an trait
}
\details{
Requires authentication
}

