% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_dataset.R
\name{POST_dataset}
\alias{POST_dataset}
\title{POST data into the Mangal datasets table}
\usage{
POST_dataset()
}
\value{
The status of the injection:
'dataset' already in mangal' means that the dataset name already have an id
'dataset done' an id has been created and the injection is succesfull
}
\description{
GET foreign keys needed for the 'datasets' table then POST
   the metadata associated. 'users' and 'refs' tables must be POST
   before.
}
\author{
Gabriel Bergeron
}
\keyword{database}
