% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_interaction.R
\name{POST_interaction}
\alias{POST_interaction}
\title{POST data in the Mangal interactions table}
\usage{
POST_interaction(inter_df = data, inter = inter, enviro = NA,
  attr = NULL, users = users)
}
\arguments{
\item{inter_df}{A dataframe with three columns:\cr
'taxon_1' and 'taxon_2': names of the taxa as they appear in the taxa table\cr
'value': value of the measured interaction\cr}

\item{inter}{A list of the interaction's metadata; must have these levels:\cr
'taxon_1_level' and 'taxon_2_level': taxonomic level of the taxa ('taxon', 'population' or 'individual')\cr
'date': YYYY-MM-DD\cr
'direction': direction of the interaction ('directed', 'undirected' of 'unknown')\cr
'type': type of biological interaction ('competition', 'amensalism', 'neutralism', 'commensalism', 'mutualism', 'parasitism', 'predation', 'herbivory', 'symbiosis', 'scavenger', 'unknown')\cr
'method': how the interaction was recorded\cr
'value': value of the interaction
'lat', 'lon' and 'srid': spacial coordinates and spacial reference id (SRID) of the interaction\cr
'public': boolean\cr}

\item{enviro}{A list of the environement's metadata; must have the levels 'name', 'date' and 'value'}

\item{attr}{A list of the interaction's attribute; must have the levels 'name' and 'unit'}

\item{users}{A list of the user's metadata; must have the level 'name'}
}
\value{
The status of the injection:
'interactions done' an id has been created for each interactions and the
injection is succesfull
}
\description{
GET foreign keys needed for the 'interactions' table then POST
   the metadata associated. 'attributes', 'environments', 'networks', 'refs'
   and 'users' tables must be POST before.
}
\author{
Gabriel Bergeron
}
\keyword{database}
