% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_line.R
\name{POST_line}
\alias{POST_line}
\title{POST a line of json data into a Mangal table}
\usage{
POST_line(table_lst_line, table)
}
\arguments{
\item{table_lst_line}{A json line of data}

\item{table}{A element, must be the name of the targeted table with ""}
}
\value{
The status of the request
}
\description{
POST a line of json data into the Mangal table specified by the user
}
\author{
Gabriel Bergeron
}
\keyword{database}
