% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_networks.R
\name{POST_networks}
\alias{POST_networks}
\title{POST data into the Mangal networks table}
\usage{
POST_networks(networks_lst)
}
\arguments{
\item{networks_lst}{A list with the network metadata}
}
\value{
The status of the injection:
'network already in mangal' means that the environment name already have an
id
'network done' an id has been created and the injection is succesfull
}
\description{
GET foreign keys needed for the 'networks' table then POST the
 metadata associated. 'environments', 'users', 'datasets' and 'refs' tables
 must be POST before.
}
\author{
Gabriel Bergeron
}
\keyword{database}
