% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_refs.R
\name{POST_refs}
\alias{POST_refs}
\title{POST data into the Mangal refs table}
\usage{
POST_refs()
}
\value{
The status of the injection:
'ref already in mangal' means that the ref name already have an id
'ref done' an id has been created and the injection is succesfull
}
\description{
POST the metadata associated with the ref table.
}
\author{
Gabriel Bergeron
}
\keyword{database}
