% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_table.R
\name{POST_table}
\alias{POST_table}
\title{POST many lines of json data on a same Mangal table}
\usage{
POST_table(data_lst, table)
}
\arguments{
\item{data_lst}{A list with data to be injected, must be json}

\item{table}{A element, must be the name of the targeted table with ""}
}
\value{
The status and line of failed attempt; if none -> empty
}
\description{
POST a list of json data in the Mangal table specified by the user
}
\author{
Gabriel Bergeron
}
\keyword{database}
