% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_taxon.R
\name{POST_taxon}
\alias{POST_taxon}
\title{POST data into the Mangal taxons table}
\usage{
POST_taxon(taxon_df = taxon_df)
}
\value{
The status of the injection:
'taxon done' ids have been created and the injection is succesfull
}
\description{
GET foreign keys needed for the 'taxons' table then POST
   the metadata associated.
}
\author{
Gabriel Bergeron
}
\keyword{database}
