% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_user.R
\name{POST_user}
\alias{POST_user}
\title{POST data into the Mangal users table}
\usage{
POST_user()
}
\value{
The status of the injection:
'user already in mangal' means that the environment name already have an id
'user done' an id has been created and the injection is succesfull
}
\description{
POST the metadata associated with the users table.
}
\author{
Gabriel Bergeron
}
\keyword{database}
