% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUT_taxa_id.R
\name{PUT_taxa_id}
\alias{PUT_taxa_id}
\title{Update taxon_id in the table taxons}
\usage{
PUT_taxa_id()
}
\description{
GET the foreign key 'taxon_id' from the table taxo_backs and update it in the table taxons
}
\author{
Gabriel Bergeron
}
\keyword{database}
