% Generated by roxygen2 (4.0.1): do not edit by hand
\name{releaseResource}
\alias{releaseResource}
\title{Release a resource}
\usage{
releaseResource(api, type, id)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{type}{the type of object to release}

\item{id}{the id of the dataset to release}
}
\description{
This function is used internally to release a resource. It do not returns
anything, but throws a warning if there is no public field to update.
}

