% Generated by roxygen2 (4.0.1): do not edit by hand
\name{render_parameters}
\alias{render_parameters}
\title{Render url additional key/value pairs}
\usage{
render_parameters(api, suppl = NULL)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{suppl}{the additional parameters as a list (with names)}
}
\description{
Render url additional key/value pairs
}

