\name{resToURI}
\alias{resToURI}
\title{Convert resource or ID to URI}
\usage{
  resToURI(api, obj, type)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{obj}{the object to convert}

  \item{obj}{the type of object to convert}
}
\description{
  Returns the URI of a resource, for internal use only
}

