% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangal-package.R
\docType{package}
\name{rmangal-package}
\alias{rmangal}
\alias{rmangal-package}
\title{rmangal: rmangal - Explore the ecological interactions database}
\description{
Retrieve and explore data from the ecological interaction database MANGAL.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://poisotlab.biol.umontreal.ca/#/}
  \item \url{https://github.com/mangal-wg/rmangal}
  \item Report bugs at \url{https://github.com/mangal-wg/rmangal/issues}
}

}
\author{
\strong{Maintainer}: Steve Vissault \email{steve.vissault@usherbrooke.ca} (0000-0002-0866-4376)

Authors:
\itemize{
  \item Gabriel Bergeron \email{gabriel.bergeron3@usherbrooke.ca} [contributor]
  \item Benjamin Mercier \email{Benjamin.B.Mercier@usherbrooke.ca} [contributor]
  \item Clément Violet \email{Clement.Violet@etudiant.univ-brest.fr} [contributor]
  \item Kevin Cazelles \email{kcazelle@uoguelph.ca} (0000-0001-6619-9874) [contributor]
  \item Timothée Poisot \email{timothee.poisot@umontreal.ca} [contributor]
}

}
\keyword{internal}
