% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rmangal_endpoints}
\alias{rmangal_endpoints}
\title{Mangal API's endpoints}
\format{
A data frame with 2 columns and 1 row per endpoint.
}
\usage{
rmangal_endpoints
}
\description{
Mangal API's endpoints
}
\references{
\itemize{
\item \url{https://mangal-interactions.github.io/mangal-api/}
}
}
\keyword{datasets}
