\name{sayHi}
\alias{sayHi}
\title{Says Hi! to the API}
\usage{
  sayHi(API)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}
}
\description{
  This function tries to connect to a mangal API. If this
  is a success, it returns the response for the API top
  level adress (giving, most notably, the API version
  number and codename).
}

