% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_datasets.R
\name{search_datasets}
\alias{search_datasets}
\title{Search mangal datasets}
\usage{
search_datasets(query = NULL, ...)
}
\arguments{
\item{query}{a \code{character} string containing a keyword used to search (case sensitive), or \code{list}of custom query (see examples).}

\item{...}{further arguments to be passed to \code{\link[rmangal:get_gen]{rmangal::get_gen()}}.}
}
\value{
object \code{data.frame}: Datasets with all networks and the original reference attached.
}
\description{
Search mangal datasets
}
\examples{
\dontrun{
all_datasets <- search_datasets()
all_datasets
class(all_datasets)
}
# Search with keyword
search_datasets(query = 'lagoon')
search_datasets(query = '2011')
# Search with custom query (specific column)
search_datasets(query = list(name = 'kemp_1977'))
}
