% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_interactions.R
\name{search_interactions}
\alias{search_interactions}
\title{Search for specific interactions type (e.g. mutualism)}
\usage{
search_interactions(type = NULL)
}
\arguments{
\item{type}{\code{character} one of the interactions type documented in the API}
}
\value{
object \code{data.frame}: Taxa information with all interactions in which the taxa occurs.
}
\description{
Search for specific interactions type (e.g. mutualism)
}
