% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_networks.R
\name{search_networks}
\alias{search_networks}
\title{Search mangal networks}
\usage{
search_networks(query = NULL, verbose = TRUE, ...)
}
\arguments{
\item{query}{\code{character} keyword used to search (case sensitive) or a \code{sf} object used to search in a specific geographical area.}

\item{verbose}{a logical. Should extra information be reported on progress?}

\item{...}{arguments from \code{\link[rmangal:get_gen]{rmangal::get_gen()}}, ignored for spatial query.}
}
\value{
\code{data.frame} object with all networks informations
Class returned \code{mgSearchNetworks}
}
\description{
Search mangal networks
}
\examples{
search_networks(query="insect\%")
\dontrun{
# Spatial query
library(USAboundaries)
area <- us_states(state="california")
networks_in_area <- search_networks(area)
plot(networks_in_area)
}
}
