% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_nodes.R
\name{search_nodes}
\alias{search_nodes}
\title{Search network by querying the nodes database table.}
\usage{
search_nodes(query, verbose = TRUE, ...)
}
\arguments{
\item{query}{either a character string including a single keyword or a list containing a custom query (see details section below).
Note that if an empty character string is passed, then all datasets available are returned.}

\item{verbose}{a \code{logical}. Should extra information be reported on progress?}

\item{...}{further arguments to be passed to \code{\link[httr:GET]{httr::GET()}}}
}
\value{
An object of class \code{mgSearchNodes}, which is a \code{data.frame} including taxa #' matching the query and corresponding information.
All networks in which taxa are involved are also attached to the \code{data.frame}.
}
\description{
Search network by querying the nodes database table.
}
\details{
If \code{query} is a character string, then all fields of the database table
including character strings are searched and entries for which at least one
partial match was found are returned.
Alternatively, a named list can be used to look for an exact match in a specific field.
In this case, the name of the list should match one of the field names of the database table.
For the \code{networks} table, those are:
Note that for lists with more than one element, only the first element is used, the others are ignored. An example is provided below.
}
\examples{
res_acer <- search_nodes("Acer")
res_926 <- search_nodes(list(network_id = 926))

}
\references{
\url{https://mangal-wg.github.io/mangal-api/#nodes}
}
\seealso{
\code{\link[=search_taxa]{search_taxa()}}
}
