% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_reference.R
\name{search_reference}
\alias{search_reference}
\title{Search for a specific reference with DOI}
\usage{
search_reference(doi = NULL)
}
\arguments{
\item{doi}{\code{character} Digital Object Identifier of the article which the network is attached.}
}
\value{
\code{data.frame} object with the reference corresponding to the query. All datasets and networks related to the publication are also attached to the \code{data.frame}.
Class returned \code{mgSearchReference}
}
\description{
Search for a specific reference with DOI
}
\examples{
search_reference(doi = "10.2307/3225248")
}
