% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgGetResponses.R
\name{summary.mgGetResponses}
\alias{summary.mgGetResponses}
\title{Summarize response status}
\usage{
\method{summary}{mgGetResponses}(object, ...)
}
\arguments{
\item{object}{object of class \code{mgGetResponses}.}

\item{...}{ignored}
}
\description{
Summarize response status
}
\keyword{internal}
