\name{toCheddar}
\alias{toCheddar}
\title{Export a network to cheddar}
\usage{
toCheddar(api, id, level = "taxa", ...)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{id}{the id of the network}

  \item{level}{the level of aggregation}

  \item{...}{reserved for future options}
}
\description{
...
}

