% Generated by roxygen2 (4.0.1): do not edit by hand
\name{uri_from_uname}
\alias{uri_from_uname}
\title{Get the URI of a user from its username}
\usage{
uri_from_uname(api, uname)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{uname}{a username}
}
\description{
Get the URI of a user from its username
}

