% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgGetResponses.R
\name{as.data.frame.mgGetResponses}
\alias{as.data.frame.mgGetResponses}
\title{Coerce \code{mgGetResponses} to data.frame}
\usage{
\method{as.data.frame}{mgGetResponses}(x, ...)
}
\arguments{
\item{x}{object of class \code{mgGetResponses}.}

\item{...}{ignored}
}
\description{
Coerce \code{mgGetResponses} to data.frame
}
