% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.igraph.R
\name{as.igraph}
\alias{as.igraph}
\alias{as.igraph.mgNetworksCollection}
\alias{as.igraph.mgNetwork}
\title{Coerce \code{mgNetworksCollection} or \code{mgNetwork} objects to \code{igraph} objects}
\usage{
as.igraph(x)

\method{as.igraph}{mgNetworksCollection}(x)

\method{as.igraph}{mgNetwork}(x)
}
\arguments{
\item{x}{\code{mgNetworksCollection} or \code{mgNetwork} object}
}
\value{
An \code{igraph} object returned by \code{\link[igraph:graph_from_data_frame]{igraph::graph_from_data_frame()}}.
}
\description{
Coerce \code{mgNetworksCollection} or \code{mgNetwork} objects to \code{igraph} objects
}
\section{Methods (by class)}{
\itemize{
\item \code{mgNetworksCollection}: Coerce \code{mgNetworksCollection} to \code{igraph} object.

\item \code{mgNetwork}: Coerce \code{mgNetwork} to \code{igraph} object.
}}

\examples{
\dontrun{
insects_networks <- get_collection(search_networks(query='insect\%'))
# Apply as.igraph on one specific network
insects_network <- insects_networks[[1]]
ig_network <- as.igraph(insects_network)
# Apply as.igraph on networks collection
ig_coll_networks <- as.igraph(insects_networks)
# Plot igraph object with vertex label
plot(ig_network)
plot(ig_network, vertex.label = insects_network$nodes$taxonomy.name)
}

}
