% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache_rmangal}
\alias{clear_cache_rmangal}
\title{Clear memoise cache}
\usage{
clear_cache_rmangal()
}
\value{
\code{TRUE} when the cache has been reset.
}
\description{
Resets the cache of the memoised function used for http GET queries (see \code{\link[memoise:forget]{memoise::forget()}}).
}
\examples{
clear_cache_rmangal()
}
