% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_citations.R
\name{get_citation}
\alias{get_citation}
\alias{get_citation.mgNetwork}
\alias{get_citation.mgNetworksCollection}
\title{Display BibTeX for all publications involved in the networks collection}
\usage{
get_citation(x)

\method{get_citation}{mgNetwork}(x)

\method{get_citation}{mgNetworksCollection}(x)
}
\arguments{
\item{x}{\code{mgNetworksCollection} or \code{mgNetworks} object class.}
}
\value{
\code{list} of bibtex reference part of the network collection
}
\description{
Display BibTeX for all publications involved in the networks collection
}
\section{Methods (by class)}{
\itemize{
\item \code{mgNetwork}: Display the BibTeX of the network's publication

\item \code{mgNetworksCollection}: Display the BibTeX of all publications part of the networks collection
}}

\examples{
 # On network collection
 lagoon_net_collection <- get_collection(search_datasets("lagoon"))
 get_citation(lagoon_net_collection)
 # On individual network
 network <- get_network_by_id(18)
 get_citation(network)
}
