% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_citations.R
\name{get_citations}
\alias{get_citations}
\title{Display all publications (BibTeX) involved in the networks collection}
\usage{
get_citations(collection)
}
\arguments{
\item{collection}{\code{mgNetworksCollection} object class.}
}
\value{
\code{list} of bibtex for all references attached to the network collection
}
\description{
Display all publications (BibTeX) involved in the networks collection
}
\examples{
lagoon_datasets <- search_datasets("lagoon")
}
