% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collection.R
\name{get_collection}
\alias{get_collection}
\alias{get_collection.default}
\alias{get_collection.mgSearchDatasets}
\alias{get_collection.mgSearchNetworks}
\alias{get_collection.mgSearchReference}
\alias{get_collection.mgSearchTaxa}
\alias{get_collection.mgSearchInteractions}
\title{Get a collection of networks}
\usage{
get_collection(x, ...)

\method{get_collection}{default}(x, ...)

\method{get_collection}{mgSearchDatasets}(x, ...)

\method{get_collection}{mgSearchNetworks}(x, ...)

\method{get_collection}{mgSearchReference}(x, ...)

\method{get_collection}{mgSearchTaxa}(x, ...)

\method{get_collection}{mgSearchInteractions}(x, ...)
}
\arguments{
\item{x}{\code{numeric} vector of mangal network IDs or object return by functions \code{search_*}.}

\item{...}{arguments to be passed on to \code{\link[rmangal:get_network_by_id]{rmangal::get_network_by_id()}}.}
}
\value{
object \code{mgNetworksCollection}:
\itemize{
\item collection of networks: \code{list} of object \code{mgNetwork} (see \code{\link[rmangal:get_network_by_id]{rmangal::get_network_by_id()}})
}
}
\description{
Get a collection of networks
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Get a collection of networks (default).

\item \code{mgSearchDatasets}: Get a collection of networks from a \code{mgSearchDatasets} object.

\item \code{mgSearchNetworks}: Get a collection of networks from a \code{mgSearchNetworks} object.

\item \code{mgSearchReference}: Get a collection of networks from a \code{mgSearchReference} object.

\item \code{mgSearchTaxa}: Get a collection of networks from a \code{mgSearchTaxa} object.

\item \code{mgSearchInteractions}: Get a collection of networks from a \code{mgSearchTaxa} object.
}}

\examples{
\dontrun{
get_collection(c(1035:1036))
get_collection(search_networks(query='insect\%'))
get_collection(search_datasets(query='lagoon\%'))
}
}
