% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collection.R
\name{get_collection}
\alias{get_collection}
\alias{get_collection.default}
\alias{get_collection.mgSearchDatasets}
\alias{get_collection.mgSearchNetworks}
\alias{get_collection.mgSearchReference}
\alias{get_collection.mgSearchNodes}
\alias{get_collection.mgSearchTaxonomy}
\alias{get_collection.mgSearchInteractions}
\title{Get a collection of networks}
\usage{
get_collection(x, ...)

\method{get_collection}{default}(x, ...)

\method{get_collection}{mgSearchDatasets}(x, ...)

\method{get_collection}{mgSearchNetworks}(x, ...)

\method{get_collection}{mgSearchReference}(x, ...)

\method{get_collection}{mgSearchNodes}(x, ...)

\method{get_collection}{mgSearchTaxonomy}(x, ...)

\method{get_collection}{mgSearchInteractions}(x, ...)
}
\arguments{
\item{x}{\code{numeric} vector of mangal network IDs or an object returned by
by on of the \code{search_*()} functions.}

\item{...}{arguments to be passed on to \code{\link[rmangal:get_network_by_id]{rmangal::get_network_by_id()}}.}
}
\value{
If there is only one network to be retrieved, \code{get_collection()} returns a
\code{mgNetwork} object, otherwise it returns a object of class
\code{mgNetworksCollection} which is a collection (a list) of \code{mgNetwork}
objects \code{\link[rmangal:get_network_by_id]{rmangal::get_network_by_id()}}).
}
\description{
Get a collection of networks
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Get a collection of networks (default).

\item \code{mgSearchDatasets}: Get a collection of networks from a \code{mgSearchDatasets} object.

\item \code{mgSearchNetworks}: Get a collection of networks from a \code{mgSearchNetworks} object.

\item \code{mgSearchReference}: Get a collection of networks from a \code{mgSearchReference} object.

\item \code{mgSearchNodes}: Get a collection of networks from a \code{mgSearchNodes} object.

\item \code{mgSearchTaxonomy}: Get a collection of networks from a \code{mgSearchTaxa} object.

\item \code{mgSearchInteractions}: Get a collection of networks from a \code{mgSearchTaxa} object.
}}

\examples{
\dontrun{
get_collection(c(1035:1036))
get_collection(search_networks(query='insect\%'))
get_collection(search_datasets(query='lagoon\%'))
}
}
