% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_fkey}
\alias{get_fkey}
\title{GET entries based on foreign key}
\usage{
get_fkey(endpoint, column, id, ...)
}
\arguments{
\item{endpoint}{\code{character} API entry point}

\item{column}{\code{character} column which contains the fkey}

\item{id}{\code{numeric} foreign key}

\item{...}{get_gen options, see \code{\link[rmangal:get_gen]{rmangal::get_gen()}}}
}
\value{
Object returned by \code{\link[rmangal:get_gen]{rmangal::get_gen()}}
}
\description{
GET entries based on foreign key
}
\details{
See endpoints available with \code{print(endpoints)}
}
