% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_gen}
\alias{get_gen}
\title{GET generic function}
\usage{
get_gen(endpoint = NULL, query = NULL, limit = 100, flatten = TRUE,
  output = "data.frame", ...)
}
\arguments{
\item{endpoint}{\code{character} API entry point}

\item{query}{\code{list} list of params passed to the API}

\item{limit}{\code{integer} number of entries return by the API (max: 1000)}

\item{flatten}{\code{logical} flatten nested data.frame, see \code{\link[jsonlite:flatten]{jsonlite::flatten()}}; default: \code{TRUE}}

\item{output}{\code{character} output type (\code{data.frame}, \code{list}, \code{spatial}, \code{raw}) return (default: data.frame)}

\item{...}{httr options, see \code{\link[httr:GET]{httr::GET()}}}
}
\value{
Object of class \code{mgGetResponses} whithin each level is a page.
Each item of the list \code{mgGetResponses} corresponds to an API call. Each call returns an object:
\itemize{
\item \code{getSuccess} which is a list with the body \code{\link[httr:content]{httr::content()}} and the server response \code{\link[httr:response]{httr::response()}}.
\item \code{getError} which has the exact same structure with an empty body.
}
}
\description{
GET generic function
}
\details{
See endpoints available with \code{print(endpoints)}
}
