% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_gen}
\alias{get_gen}
\title{Generic API function to retrieve several entries}
\usage{
get_gen(endpoint, query = NULL, limit = 100, verbose = TRUE, ...)
}
\arguments{
\item{endpoint}{\code{character} API entry point}

\item{query}{\code{list} list of parameters passed to the API}

\item{limit}{\code{integer} number of entries return by the API (max: 1000)}

\item{verbose}{\code{logical} print API code status on error; default: \code{TRUE}}

\item{...}{Further named parameters passed to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
Object of class \code{mgGetResponses}
}
\description{
Generic API function to retrieve several entries
}
\details{
See endpoints available with \code{endpoints()}
}
\keyword{internal}
