% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_by_id.R
\name{get_network_by_id}
\alias{get_network_by_id}
\title{Retrieve mangal network by id}
\usage{
get_network_by_id(id, ...)
}
\arguments{
\item{id}{\code{numeric} mangal ID network}

\item{...}{arguments from \code{\link[rmangal:get_singletons]{rmangal::get_singletons()}}}
}
\value{
a \code{mgNetwork} object which include:
\itemize{
\item network: \code{list} of all generic informations on the network
\item nodes: \code{data.frame} of all nodes with taxonomic informations
\item edges: \code{data.frame} of all edges (ecological interactions), with the attribute used to describe the interaction
\item dataset: \code{list} of all informations on the dataset associated to the network
\item reference: \code{list} of all informations on the original publication
}
}
\description{
Retrieve mangal network by id
}
\examples{
 get_network_by_id(id = 18)
}
