% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_by_id.R
\name{get_network_by_id}
\alias{get_network_by_id}
\title{Retrieve mangal network by id}
\usage{
get_network_by_id(id, ...)
}
\arguments{
\item{id}{\code{numeric} mangal ID network}

\item{...}{arguments from \code{\link[rmangal:get_singletons]{rmangal::get_singletons()}}}
}
\value{
a \code{mgNetwork} object including:
\itemize{
\item network: a \code{list} of all generic informations on the network;
\item nodes: a \code{data.frame} of all nodes with taxonomic informations;
\item edges: a \code{data.frame} of all edges (ecological interactions), with the attribute used to describe the interaction
\item dataset: \code{list} information pertaining to the datasetthe network is associaated to;
\item reference: \code{list} information about the original publication.
}
}
\description{
Retrieve mangal network by id
}
\examples{
 get_network_by_id(id = 18)
}
