% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_by_id.R
\name{get_network_by_id}
\alias{get_network_by_id}
\alias{get_network_by_id_indiv}
\alias{print.mgNetwork}
\alias{print.mgNetworksCollection}
\title{Retrieve network data for a given set of Mangal identifiera #'}
\usage{
get_network_by_id(ids, verbose = TRUE)

get_network_by_id_indiv(id, verbose = TRUE)

\method{print}{mgNetwork}(x, ...)

\method{print}{mgNetworksCollection}(x, ...)
}
\arguments{
\item{ids}{a vector of Mangal ID for networks (\code{numeric}).}

\item{verbose}{a logical. Should extra information be reported on progress?}

\item{id}{a single ID network (\code{numeric}).}

\item{x}{an object of class \code{mgNetwork} or \code{mgNetworksCollection}.}

\item{...}{ignored.}
}
\value{
a \code{mgNetwork} object including:
\itemize{
\item network: a \code{list} of all generic information on the network;
\item nodes: a \code{data.frame} of all nodes with taxonomic information;
\item edges: a \code{data.frame} of all edges (ecological interactions), with the attribute used to describe the interaction #' - dataset: \code{list} information pertaining to the dataset the network is associated with;
\item reference: \code{list} information about the original publication.
}
}
\description{
Retrieve network data for a given set of Mangal identifiera #'
}
\section{Functions}{
\itemize{
\item \code{get_network_by_id_indiv}: Retrieve a network by its  collection of networks (default).
}}

\examples{
net18 <- get_network_by_id(id = 18) 
nets <- get_network_by_id(id = c(18, 23))
}
