% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_singletons}
\alias{get_singletons}
\title{Generic API function to retrieve singletons}
\usage{
get_singletons(endpoint = NULL, ids = NULL, output = "data.frame",
  flatten = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{endpoint}{\code{character} API entry point}

\item{ids}{\code{numeric} vector of ids}

\item{output}{\code{character} output type (\code{data.frame}, \code{list}, \code{spatial}, \code{raw}) return; default: \code{list}}

\item{flatten}{\code{logical} return flatten nested data.frame, see \code{\link[jsonlite:flatten]{jsonlite::flatten()}}; default: \code{TRUE}}

\item{verbose}{\code{logical} print API code status on error; default: \code{TRUE}}

\item{...}{httr options, see \code{\link[httr:GET]{httr::GET()}}}
}
\value{
Object of class \code{mgGetResponses} whithin each level is the specific ID called.
Each item of the \code{mgGetResponses} list corresponds to an API call. Each call returns an object:
\itemize{
\item \code{getSuccess} which is a list with the body \code{\link[httr:content]{httr::content()}} and the server response \code{\link[httr:response]{httr::response()}}.
\item \code{getError} which has the exact same structure with an empty body.
}
}
\description{
Generic API function to retrieve singletons
}
\details{
See endpoints available with \code{endpoints()}
}
\keyword{internal}
