% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_singletons}
\alias{get_singletons}
\title{Generic API function to retrieve singletons}
\usage{
get_singletons(endpoint = NULL, ids = NULL, verbose = TRUE, ...)
}
\arguments{
\item{endpoint}{\code{character} API entry point.}

\item{ids}{\code{numeric} vector of ids.}

\item{verbose}{\code{logical} print API code status on error; default: \code{TRUE}}

\item{...}{httr options, see \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
Object of class \code{mgGetResponses}
}
\description{
Generic API function to retrieve singletons
}
\details{
See endpoints available with \code{endpoints()}
}
\keyword{internal}
