\name{getPopstate}
\alias{getPopstate}
\title{Get a particular population state}
\usage{
  getPopstate(API, id)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{id}{The ID of the population state to retrieve}
}
\description{
  Returns the state of a given population
}

