% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{is.mgGetResponses}
\alias{is.mgGetResponses}
\title{Test class \code{coleoGetResponses}}
\usage{
is.mgGetResponses(x)
}
\arguments{
\item{x}{R object to test}
}
\description{
Test class \code{coleoGetResponses}
}
