% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgGetResponses.R
\name{is.mgGetResponses}
\alias{is.mgGetResponses}
\title{Test class \code{coleoGetResponses}}
\usage{
is.mgGetResponses(x)
}
\arguments{
\item{x}{R x to test}
}
\description{
Test class \code{coleoGetResponses}
}
