% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List mangal datasets with networks and reference attached}
\usage{
list_datasets(search = NULL, ...)
}
\arguments{
\item{search}{\code{character} keyword to search (case sensitive).}

\item{...}{arguments from \code{\link[rmangal:get_gen]{rmangal::get_gen()}}}
}
\value{
object \code{data.frame}: Datasets with all networks and the original reference attached.
}
\description{
List mangal datasets with networks and reference attached
}
\details{
See endpoints available with \code{print(endpoints)}
With \code{search} argument, the '%' can be used to represent any character or set of characters before or after the keyword (e.g '%lagoon%')
}
\examples{
list_datasets()
list_datasets(search="\%lagoon\%")
}
