\name{listNetworks}
\alias{listNetworks}
\title{Lists the networks that are part of a dataset}
\usage{
  listNetworks(API, dataset, network)

  listNetworks(API, dataset, network)
}
\arguments{
  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{dataset}{A \code{list} object representing a
  dataset}

  \item{API}{An object of class \code{\link{mangalAPI}}}

  \item{dataset}{A \code{list} object representing a
  dataset}

  \item{network}{A \code{list} object representing a
  network}
}
\description{
  Makes a list of all networks part of a dataset

  Makes a list of all interactions part of a network
}

