\name{mangalapi}
\alias{mangalapi}
\title{Initialise a mangal API object}
\usage{
  mangalapi(url = "http://mangal.uqar.ca", v = "v1",
    usr = NULL, pwd = NULL)
}
\arguments{
  \item{url}{The URL to the server}

  \item{v}{The API version}

  \item{usr}{The username}

  \item{pwd}{The password}
}
\description{
  Initialise a mangal API obect, with direct reference to
  the resources
}

