\name{mangalapi}
\alias{mangalapi}
\title{Initialise a mangal API object}
\usage{
mangalapi(url = "http://mangal.uqar.ca", v = "v1", usr = NULL,
  key = NULL)
}
\arguments{
  \item{url}{The URL to the server}

  \item{v}{The API version}

  \item{usr}{The username}

  \item{pwd}{The API key}
}
\description{
Initialise a mangal API obect, with direct reference to the
resources
}

