% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_network.R
\name{mg_network}
\alias{mg_network}
\title{Retrieve full network (edges and nodes)}
\usage{
mg_network(id = NULL, ...)
}
\arguments{
\item{id}{\code{numeric} mangal ID network}
}
\value{
object `mgNetwork:
\itemize{
\item network: \code{list} of all generic informations on the network
\item nodes: \code{data.frame} of all nodes with taxonomic informations
\item edges: \code{data.frame} of all edges (ecological interactions), with the attribute used to describe the interaction
}
}
\description{
Retrieve full network (edges and nodes)
}
\examples{
mg_network(id = 18)
}
