\name{multi_resToURI}
\alias{multi_resToURI}
\title{Convert several resources or IDs to URI}
\usage{
multi_resToURI(api, obj, type)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{obj}{the object to convert}

  \item{type}{the type of object to convert}
}
\description{
For internal use only
}

