% Generated by roxygen2 (4.0.1): do not edit by hand
\name{patchInteraction}
\alias{patchInteraction}
\title{Patch an interaction}
\usage{
patchInteraction(api, data)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{data}{the interaction in list format}
}
\description{
Patch an interaction from the database
}
\details{
Requires authentication.

If \code{item_*} or \code{pop_*} are provided in \code{data}, it is not necessary to fill-in other top-level fields (\code{taxa_*} and \code{pop_*}). If they are not filled, they will be retrieved by the package (at the cost of a few additional queries, which might rapidly add-up for large datasets).
}

