\name{patchPopulation}
\alias{patchPopulation}
\title{Patch a population}
\usage{
patchPopulation(api, data)

patchPopulation(api, data)
}
\arguments{
  \item{api}{a \code{\link{mangalapi}} object}

  \item{data}{the population in list format}

  \item{api}{a \code{\link{mangalapi}} object}

  \item{data}{the reference in list format}
}
\description{
Modify the informations for a population

Modify the informations for a reference
}
\details{
Requires authentication

Requires authentication
}

