% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_attribute.R
\name{POST_attribute}
\alias{POST_attribute}
\title{POST data into the Mangal attributes table}
\usage{
POST_attribute(data)
}
\arguments{
\item{attr}{A list of four levels: name, table_owner, description and unit}
}
\value{
The status of the injection:
'attr already in mangal' means that the attribute name already have an id
'attr done' an id has been created and the injection is succesfull
}
\description{
POST the metadata associated whit the attributes table
}
\author{
Gabriel Bergeron
}
\keyword{database}
