% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_environment.R
\name{POST_environment}
\alias{POST_environment}
\title{POST data into the Mangal environments table}
\usage{
POST_environment(enviro = enviro, attr = attr)
}
\value{
The status of the injection:
'enviro already in mangal' means that the environment name already have an id
'enviro done' an id has been created and the injection is succesfull
}
\description{
GET foreign keys needed for the 'environments' table then POST
   the metadata associated. 'attributes' and 'refs' tables must be POST
   before.
}
\author{
Gabriel Bergeron
}
\keyword{database}
