% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_interaction.R
\name{POST_interaction}
\alias{POST_interaction}
\title{POST data in the Mangal interactions table}
\usage{
POST_interaction(inter_df = data, enviro = NA, attr = NA)
}
\arguments{
\item{inter_df}{A dataframe with three columns: taxon_1, taxon_2 and value}
}
\value{
The status of the injection:
'interactions done' an id has been created for each interactions and the
injection is succesfull
}
\description{
GET foreign keys needed for the 'interactions' table then POST
   the metadata associated. 'attributes', 'environments', 'networks', 'refs'
   and 'users' tables must be POST before.
}
\author{
Gabriel Bergeron
}
\keyword{database}
