% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_ref.R
\name{POST_ref}
\alias{POST_ref}
\title{POST data into the Mangal refs table}
\usage{
POST_ref(ref = ref)
}
\arguments{
\item{ref}{A list of the reference's metadata; must have these levels:\cr
'doi': DOI of the attached publication\cr
'jstor': JSTOR of the attached publication\cr
'pmid': PMID of the attached publication\cr
'paper_url': URL of the attached publication\cr
'data_url': URL of the attached data\cr
'author': first author name\cr
'year': year of publication\cr
'bibtex': BibTex of the attached publication\cr}
}
\value{
The status of the injection:
'ref already in mangal' means that the ref name already have an id
'ref done' an id has been created and the injection is succesfull
}
\description{
POST the metadata associated with the ref table.
}
\author{
Gabriel Bergeron
}
\keyword{database}
