% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_taxa_back.R
\name{POST_taxa_back}
\alias{POST_taxa_back}
\title{POST data into the Mangal taxo_back table}
\usage{
POST_taxa_back()
}
\value{
The status of the injection:
'taxo_back done' ids have been created and the injection is succesfull
}
\description{
POST the metadata associated to the taxonomic backbone.
}
\author{
Gabriel Bergeron
}
\keyword{database}
