% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_taxon.R
\name{POST_taxon}
\alias{POST_taxon}
\title{POST data into the Mangal taxons table}
\usage{
POST_taxon(taxon_df = taxon_df)
}
\arguments{
\item{taxon_df}{A dataframe of two columns:\cr
'original_name': name of the taxa as found in the publication\cr
'name_clear': clean taxonomy, name of the taxa without numbers, punctuations or 'sp'}
}
\value{
The status of the injection:
'taxon done' ids have been created and the injection is succesfull
}
\description{
GET foreign keys needed for the 'taxons' table then POST
   the metadata associated.
}
\author{
Gabriel Bergeron
}
\keyword{database}
