% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_trait.R
\name{POST_trait}
\alias{POST_trait}
\title{POST data into the Mangal traits table}
\usage{
POST_trait(trait_df, network)
}
\arguments{
\item{network}{A list the network's metadata; must have the 'name' level}

\item{traits_df}{A dataframe with three columns:\cr
'taxa': taxa name as found in the taxa table
'name': name of the trait as found in the attribute table\cr
'value': value of the trait}
}
\value{
The status of the injection:
'trait already in mangal' means that the environment name already have an id
'trait done' an id has been created and the injection is succesfull
}
\description{
GET foreign keys needed for the 'traits' table then POST
   the metadata associated. 'attributes' and 'refs' tables must be POST
   before.
}
\author{
Gabriel Bergeron
}
\keyword{database}
