% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resToURI}
\alias{resToURI}
\title{Convert resource or ID to URI}
\usage{
resToURI(api, obj, type)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{obj}{the object to convert}

\item{type}{the type of object to convert}
}
\description{
Returns the URI of a resource, for internal use only
}

