% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rmangal}
\alias{rmangal}
\alias{rmangal-package}
\title{rmangal}
\description{
A programmatic interface to the Mangal API \url{https://mangal-wg.github.io/mangal-api//}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mangal.io}
  \item \url{https://github.com/ropensci/rmangal}
  \item Report bugs at \url{https://github.com/ropensci/rmangal/issues}
}

}
\author{
\strong{Maintainer}: Steve Vissault \email{steve.vissault@usherbrooke.ca} (0000-0002-0866-4376)

Authors:
\itemize{
  \item Kevin Cazelles \email{kcazelle@uoguelph.ca} (0000-0001-6619-9874) [contributor]
  \item Gabriel Bergeron \email{gabriel.bergeron3@usherbrooke.ca} [contributor]
  \item Benjamin Mercier \email{Benjamin.B.Mercier@usherbrooke.ca} [contributor]
  \item Clément Violet \email{Clement.Violet@etudiant.univ-brest.fr} [contributor]
  \item Dominique Gravel \email{dominique.gravel@usherbrooke.ca}
  \item Timothée Poisot \email{timothee.poisot@umontreal.ca}
}

Other contributors:
\itemize{
  \item Thomas Lin Pedersen (0000-0002-5147-4711) [reviewer]
  \item Anna Willoughby (0000-0002-0504-0605) [reviewer]
}

}
\keyword{internal}
