% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rmangal}
\alias{rmangal}
\alias{rmangal-package}
\title{rmangal}
\description{
A programmatic interface to the Mangal API \url{https://mangal-interactions.github.io/mangal-api/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/rmangal/}
  \item \url{https://mangal.io}
  \item \url{https://github.com/ropensci/rmangal}
  \item Report bugs at \url{https://github.com/ropensci/rmangal/issues}
}

}
\author{
\strong{Maintainer}: Kevin Cazelles \email{kevin.cazelles@gmail.com} (\href{https://orcid.org/0000-0001-6619-9874}{ORCID})

Authors:
\itemize{
  \item Steve Vissault \email{s.vissault@yahoo.fr} (\href{https://orcid.org/0000-0002-0866-4376}{ORCID}) [contributor]
  \item Gabriel Bergeron \email{gabriel.bergeron3@usherbrooke.ca} [contributor]
  \item Benjamin Mercier \email{Benjamin.B.Mercier@usherbrooke.ca} [contributor]
  \item Clément Violet \email{Clement.Violet@etudiant.univ-brest.fr} [contributor]
  \item Dominique Gravel \email{dominique.gravel@usherbrooke.ca}
  \item Timothée Poisot \email{timothee.poisot@umontreal.ca}
}

Other contributors:
\itemize{
  \item Thomas Lin Pedersen (\href{https://orcid.org/0000-0002-5147-4711}{ORCID}) [reviewer]
  \item Anna Willoughby (\href{https://orcid.org/0000-0002-0504-0605}{ORCID}) [reviewer]
}

}
\keyword{internal}
