% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_datasets.R
\name{search_datasets}
\alias{search_datasets}
\title{Search over all datasets using keyword}
\usage{
search_datasets(query = NULL, verbose = TRUE, ...)
}
\arguments{
\item{query}{a \code{character} string containing a keyword used to search (case sensitive),
or a \code{list} containiing a custom query  custom query (see examples).}

\item{verbose}{a logical. Should extra information be reported on progress?}

\item{...}{further arguments to be passed to \code{\link[rmangal:get_gen]{rmangal::get_gen()}}.}
}
\value{
An object of class \code{mgSearchDatasets}, which is a \code{data.frame}  object with all datasets corresponding to the query.
For each dataset entry, the networks and the original reference are attached.
}
\description{
Search over all datasets using keyword
}
\examples{
\dontrun{
# Return all dataset
all_datasets <- search_datasets()
all_datasets
class(all_datasets)
}
# Search with keyword
search_datasets(query = 'lagoon')
search_datasets(query = '2011')
# Search with custom query (specific column)
search_datasets(query = list(name = 'kemp_1977'))
}
