% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_interactions.R
\name{search_interactions}
\alias{search_interactions}
\title{Search for specific interactions type (e.g. mutualism)}
\usage{
search_interactions(type = avail_type(), expand_node = FALSE,
  verbose = FALSE)
}
\arguments{
\item{type}{a \code{character} one of the interactions type available in the function \code{avail_type()}.}

\item{expand_node}{a \code{logical}. Should the function returned extra information on nodes? Default is set to \code{FALSE}, which means that only the mangal ID of nodes are returned.}

\item{verbose}{a \code{logical}. Should extra information be reported on progress?}
}
\value{
An object of class \code{mgSearchInteractions}, which is a \code{data.frame} object with all interaction matching the interaction type provided.
All networks in which interactions are involved are also attached to the \code{data.frame}.
}
\description{
Search for specific interactions type (e.g. mutualism)
}
\examples{
search_interactions("competition")
# Get all networks containing competition
competition_networks <- get_collection(search_interactions("competition"))
}
\references{
\url{https://mangal-wg.github.io/mangal-api/#interactions}
}
