% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_interactions.R
\name{search_interactions}
\alias{search_interactions}
\title{Search for specific interactions type}
\usage{
search_interactions(query = NULL, ...)
}
\arguments{
\item{query}{\code{character} keyword used to search (case sensitive), or \code{list} of custom query.}
}
\value{
object \code{data.frame}: Taxa information with all interactions in which the taxa occurs.
}
\description{
Search for specific interactions type
}
\details{
A hardwork have been done to homogenized the taxonomy
}
