% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_networks.R
\name{search_networks}
\alias{search_networks}
\title{Search over all networks using keyword or spatial object}
\usage{
search_networks(query = NULL, verbose = TRUE, ...)
}
\arguments{
\item{query}{\code{character} keyword used to search over all networks (case sensitive) or a \code{sf} object
used to search in a specific geographical area. If keyword is unspecified (query = NULL), all networks will be returned.}

\item{verbose}{a \code{logical}. Should extra information be reported on progress?}

\item{...}{further arguments to be passed to \code{\link[rmangal:get_gen]{rmangal::get_gen()}}.}
}
\value{
An object of class \code{mgSearchNetworks}, which is a \code{data.frame} object with all networks informations
}
\description{
Search over all networks using keyword or spatial object
}
\examples{
search_networks(query="insect\%")
\donttest{
# Spatial query
library(USAboundaries)
area <- us_states(state="california")
networks_in_area <- search_networks(area)
plot(networks_in_area)
}
}
