% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_reference.R
\name{search_reference}
\alias{search_reference}
\title{Search for a specific dataset reference with its DOI}
\usage{
search_reference(query, doi = NULL, verbose = TRUE, ...)
}
\arguments{
\item{query}{\code{character} a Digital Object Identifier of the article (mandatory arg).}

\item{doi}{\code{character} a Digital Object Identifier of the article (mandatory arg).}

\item{verbose}{a \code{logical}. Should extra information be reported on progress?}

\item{...}{further arguments to be passed to \code{\link[rmangal:get_gen]{rmangal::get_gen()}}.}
}
\value{
An object of class \code{mgSearchReference}, which is essentially a list that includes a wide range of details associated to the reference, including all datasets and networks related to the publication that are included in mangal data base.
}
\description{
Search for a specific dataset reference with its DOI
}
\examples{
search_reference(doi = "10.2307/3225248")

}
\references{
\url{https://mangal-wg.github.io/mangal-api/#references}
}
