% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_reference.R
\name{search_reference}
\alias{search_reference}
\title{Search mangal taxons}
\usage{
search_reference(doi = NULL, pmid = NULL, jstor = NULL)
}
\arguments{
\item{doi}{\code{character} keyword used to search (case sensitive), or \code{list} of custom query.}

\item{pmid}{}
}
\value{
object \code{data.frame}: Taxa information with all interactions in which the taxa occurs.
}
\description{
Search mangal taxons
}
