% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_reference.R
\name{search_reference}
\alias{search_reference}
\title{Search for a specific dataset reference with DOI}
\usage{
search_reference(doi = NULL, verbose = TRUE, ...)
}
\arguments{
\item{doi}{\code{character} a Digital Object Identifier of the article (mandatory arg)}

\item{verbose}{a \code{logical}. Should extra information be reported on progress?}

\item{...}{further arguments to be passed to \code{\link[rmangal:get_gen]{rmangal::get_gen()}}.}
}
\value{
An object of class \code{mgSearchReference}, which is essentially a
\code{data.frame} object with the reference of the article matching the query. Note that
all datasets and networks related to the publication are also attached to the
\code{data.frame}.
}
\description{
Search for a specific dataset reference with DOI
}
\examples{
search_reference(doi = "10.2307/3225248")
}
