% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_interactions.R, R/search_taxonomy.R
\name{search_taxa}
\alias{search_taxa}
\title{Search on the taxonomy table}
\usage{
search_taxa(query = NULL)

search_taxa(query = NULL)
}
\arguments{
\item{query}{\code{character} keyword used to search (case sensitive), or \code{list} of custom query.}

\item{query}{\code{character} specific taxon / keyword / or tsn, eol, bold & ncbi IDs.}

\item{original}{a \code{logical}. Should query taxa from the original publication? (see details section)}

\item{verbose}{a \code{logical}. Should extra information be reported on progress?}
}
\value{
object \code{data.frame}: Taxa information with all interactions in which the taxa occurs.

An object of class \code{mgSearchTaxonomy}, which is a \code{data.frame} object with all taxa matching the query.
All networks in which taxa are involved are also attached to the \code{data.frame}.
}
\description{
Search on the taxonomy table

Search network by taxa
}
\details{
A hardwork have been done to homogenized the taxonomy

By default, you are querying taxa based on a homogenized taxonomy across all datasets using
TNRS (http://tnrs.iplantcollaborative.org/) and/or GNR (https://resolver.globalnames.org/).
Be aware that these taxa might not be the taxa from the original publication.
To query original documented taxa, please set the argument \code{original} as TRUE.
}
