% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_taxons.R
\name{search_taxa}
\alias{search_taxa}
\title{Search mangal taxa}
\usage{
search_taxa(query = NULL, ...)
}
\arguments{
\item{query}{\code{character} keyword used to search (case sensitive), or \code{list} of custom query.}

\item{...}{arguments from \code{\link[rmangal:get_gen]{rmangal::get_gen()}}}
}
\value{
object \code{data.frame}: Taxa information with all interactions in which the taxa occurs.
}
\description{
Search mangal taxa
}
