% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_taxa.R
\name{search_taxa}
\alias{search_taxa}
\title{Search network by taxa using keyword, tsn, eol, bold or ncbi IDs.}
\usage{
search_taxa(query = NULL, original = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{query}{\code{character} specific taxon / keyword / or tsn, eol, bold & ncbi IDs.}

\item{original}{a \code{logical}. Should query taxa names from the original publication? (see details section, default: FALSE)}

\item{verbose}{a \code{logical}. Should extra information be reported on progress?}

\item{...}{further arguments to be passed to \code{\link[rmangal:get_gen]{rmangal::get_gen()}}.}
}
\value{
An object of class \code{mgSearchTaxonomy}, which is a \code{data.frame} object with all taxa matching the query.
All networks in which taxa are involved are also attached to the \code{data.frame}.
}
\description{
Search network by taxa using keyword, tsn, eol, bold or ncbi IDs.
}
\details{
By default, you are querying taxa based on a homogenized taxonomy (corrected names from
\href{http://tnrs.iplantcollaborative.org/}{TNRS} and/or \href{https://resolver.globalnames.org/}{GNR}.
These taxa might not be the taxa recorded within the original publication.
To query original documented taxa, please set the argument \code{original} as TRUE.
}
\examples{
search_taxa("Acer")
}
