% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{summary.mgGetResponses}
\alias{summary.mgGetResponses}
\title{Summary of response status
TODO: Improve display by providing calls that are failed}
\usage{
\method{summary}{mgGetResponses}(responses)
}
\description{
Summary of response status
TODO: Improve display by providing calls that are failed
}
