% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{summary.mgGetResponses}
\alias{summary.mgGetResponses}
\title{Summary of response status}
\usage{
\method{summary}{mgGetResponses}(responses)
}
\arguments{
\item{responses}{mgGetResponses class}
}
\description{
Summary of response status
}
