% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toIgraph}
\alias{toIgraph}
\title{Export a network to igraph}
\usage{
toIgraph(api, id, level = "taxa", ...)
}
\arguments{
\item{api}{a \code{\link{mangalapi}} object}

\item{id}{the id of the network}

\item{level}{the level of aggregation}

\item{...}{reserved for future options}
}
\description{
...
}

